fix: 관리자 정산 응답 금액 타입을 BigDecimal로 변경 (#80)

+++ backend/src/main/java/com/back/team9/moyeota/domain/admin/dto/AdminSettlementListResponse.java
@@ -3,6 +3,7 @@ package com.back.team9.moyeota.domain.admin.dto;
 import com.back.team9.moyeota.domain.settlement.entity.Settlement;
 import com.back.team9.moyeota.domain.settlement.entity.SettlementStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record AdminSettlementListResponse(
@@ -11,9 +12,9 @@ public record AdminSettlementListResponse(
         String hostEmail,
         Long fundingId,
         String fundingTitle,
-        Integer totalAmount,
-        Integer platformFee,
-        Integer hostPaybackAmount,
+        BigDecimal totalAmount,
+        BigDecimal platformFee,
+        BigDecimal hostPaybackAmount,
         SettlementStatus status,
         LocalDateTime paybackPaidAt,
         LocalDateTime createdAt,
